Extract `ActiveRecord::ConnectionAdapters::OracleEnhanced::DbmsOutput`
