+++ spec/active_record/connection_adapters/oracle_enhanced/dbms_output_spec.rb
